tests/wpt/tests/css/css-backgrounds/parsing/background-position-y-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 4: getComputedStyle().backgroundPositionY</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds-4/#propdef-background-position-y">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("background-position-y", "center", "50%");
test_computed_value("background-position-y", "top", "0%");
test_computed_value("background-position-y", "bottom", "100%");
test_computed_value("background-position-y", "y-start");
test_computed_value("background-position-y", "y-end");
test_computed_value("background-position-y", "-20%");
test_computed_value("background-position-y", "10px");
test_computed_value("background-position-y", "0.5em", "20px");
test_computed_value("background-position-y", "calc(10px - 0.5em)", "-10px");
test_computed_value("background-position-y", "top -20%", "-20%");
test_computed_value("background-position-y", "bottom -10px", "calc(100% + 10px)");
test_computed_value("background-position-y", "-20%, 10px", "-20%");
test_computed_value("background-position-y", "center, top, bottom", "50%");
test_computed_value("background-position-y", "0.5em, y-start, y-end", "20px");

// See background-computed.html for a test with multiple background images.
test_computed_value("background-position-y", "calc(10px - 0.5em), -20%, 10px", "-10px");
test_computed_value("background-position-y", "calc(10px - 0.5em), top -20%, bottom 10px", "-10px");
</script>
</body>
</html>
